+++ crates/wdl-core/src/concern.rs
@@ -45,8 +45,8 @@ impl Concern {
     /// ```
     /// use wdl_core::concern::code::Kind;
     /// use wdl_core::concern::lint;
-    /// use wdl_core::concern::lint::Group;
     /// use wdl_core::concern::lint::Level;
+    /// use wdl_core::concern::lint::TagSet;
     /// use wdl_core::concern::parse;
     /// use wdl_core::concern::validation;
     /// use wdl_core::concern::Code;
@@ -72,7 +72,7 @@ impl Concern {
     /// let warning = lint::warning::Builder::default()
     ///     .code(Code::try_new(Kind::Warning, Version::V1, 1)?)
     ///     .level(Level::High)
-    ///     .group(Group::Style)
+    ///     .tags(TagSet::new(&[lint::Tag::Style]))
     ///     .push_location(Location::Unplaced)
     ///     .subject("Hello, world!")
     ///     .body("A body.")
@@ -99,8 +99,8 @@ impl Concern {
     /// ```
     /// use wdl_core::concern::code::Kind;
     /// use wdl_core::concern::lint;
-    /// use wdl_core::concern::lint::Group;
     /// use wdl_core::concern::lint::Level;
+    /// use wdl_core::concern::lint::TagSet;
     /// use wdl_core::concern::parse;
     /// use wdl_core::concern::validation;
     /// use wdl_core::concern::Code;
@@ -126,7 +126,7 @@ impl Concern {
     /// let warning = lint::warning::Builder::default()
     ///     .code(Code::try_new(Kind::Warning, Version::V1, 1)?)
     ///     .level(Level::High)
-    ///     .group(Group::Style)
+    ///     .tags(TagSet::new(&[lint::Tag::Style]))
     ///     .push_location(Location::Unplaced)
     ///     .subject("Hello, world!")
     ///     .body("A body.")
@@ -153,8 +153,8 @@ impl Concern {
     /// ```
     /// use wdl_core::concern::code::Kind;
     /// use wdl_core::concern::lint;
-    /// use wdl_core::concern::lint::Group;
     /// use wdl_core::concern::lint::Level;
+    /// use wdl_core::concern::lint::TagSet;
     /// use wdl_core::concern::parse;
     /// use wdl_core::concern::validation;
     /// use wdl_core::concern::Code;
@@ -180,7 +180,7 @@ impl Concern {
     /// let warning = lint::warning::Builder::default()
     ///     .code(Code::try_new(Kind::Warning, Version::V1, 1)?)
     ///     .level(Level::High)
-    ///     .group(Group::Style)
+    ///     .tags(TagSet::new(&[lint::Tag::Style]))
     ///     .push_location(Location::Unplaced)
     ///     .subject("Hello, world!")
     ///     .body("A body.")
@@ -207,8 +207,8 @@ impl Concern {
     /// ```
     /// use wdl_core::concern::code::Kind;
     /// use wdl_core::concern::lint;
-    /// use wdl_core::concern::lint::Group;
     /// use wdl_core::concern::lint::Level;
+    /// use wdl_core::concern::lint::TagSet;
     /// use wdl_core::concern::parse;
     /// use wdl_core::concern::validation;
     /// use wdl_core::concern::Code;
@@ -234,7 +234,7 @@ impl Concern {
     /// let warning = lint::warning::Builder::default()
     ///     .code(Code::try_new(Kind::Warning, Version::V1, 1)?)
     ///     .level(Level::High)
-    ///     .group(Group::Style)
+    ///     .tags(TagSet::new(&[lint::Tag::Style]))
     ///     .push_location(Location::Unplaced)
     ///     .subject("Hello, world!")
     ///     .body("A body.")
@@ -261,8 +261,8 @@ impl Concern {
     /// ```
     /// use wdl_core::concern::code::Kind;
     /// use wdl_core::concern::lint;
-    /// use wdl_core::concern::lint::Group;
     /// use wdl_core::concern::lint::Level;
+    /// use wdl_core::concern::lint::TagSet;
     /// use wdl_core::concern::parse;
     /// use wdl_core::concern::validation;
     /// use wdl_core::concern::Code;
@@ -288,7 +288,7 @@ impl Concern {
     /// let warning = lint::warning::Builder::default()
     ///     .code(Code::try_new(Kind::Warning, Version::V1, 1)?)
     ///     .level(Level::High)
-    ///     .group(Group::Style)
+    ///     .tags(TagSet::new(&[lint::Tag::Style]))
     ///     .push_location(Location::Unplaced)
     ///     .subject("Hello, world!")
     ///     .body("A body.")
@@ -315,8 +315,8 @@ impl Concern {
     /// ```
     /// use wdl_core::concern::code::Kind;
     /// use wdl_core::concern::lint;
-    /// use wdl_core::concern::lint::Group;
     /// use wdl_core::concern::lint::Level;
+    /// use wdl_core::concern::lint::TagSet;
     /// use wdl_core::concern::parse;
     /// use wdl_core::concern::validation;
     /// use wdl_core::concern::Code;
@@ -342,7 +342,7 @@ impl Concern {
     /// let warning = lint::warning::Builder::default()
     ///     .code(Code::try_new(Kind::Warning, Version::V1, 1)?)
     ///     .level(Level::High)
-    ///     .group(Group::Style)
+    ///     .tags(TagSet::new(&[lint::Tag::Style]))
     ///     .push_location(Location::Unplaced)
     ///     .subject("Hello, world!")
     ///     .body("A body.")