+++ crates/wdl-ast/CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Fix ignoring comments in expressions ([#23](https://github.com/stjude-rust-labs/wdl/pull/23)).
 
+### Changed
+
+* Changes the singular `Group` feature of lint warnings to one or more `Tags` (#37, contributed by @a-frantz)
+
 ## 0.1.0 — 12-17-2023
 
 ### Added