+++ crates/wdl-core/CHANGELOG.md
@@ -5,7 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## Unreleased
+
+### Added
+
+* Adds `Clarity` and `Portability` tags (#37, contributed by @a-frantz)
+
+### Changed
+
+* Changes the singular `Group` feature of lint warnings to one or more `Tags` (#37, contributed by @a-frantz)
 
 ## 0.1.0 — 12-17-2023