+++ crates/wdl-grammar/CHANGELOG.md
@@ -23,6 +23,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Adds the `one_empty_line` rule that ensures no excess of empty lines
   (#33, contributed by @simojoe).
 
+### Changed
+
+* Changes the singular `Group` feature of lint warnings to one or more `Tags` (#37, contributed by @a-frantz)
+* Changes the tags and levels of various rules to better align with #12 (#37, contributed by @a-frantz)
+
 ## 0.2.0 - 12-17-2023
 
 ### Added